requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
mistralai>=0.4.0
transformers>=4.30.0
torch>=2.0.0
spacy>=3.6.0
yake>=0.4.8
requests>=2.28.0
pydantic>=2.0.0
